:arrow_right: geohub.js -> index.js
